Move helper methods to transforms needing them
